+++ src/mkobj.c
@@ -998,7 +998,7 @@ mksobj_init(struct obj *otmp, boolean artif)
          * create an artifact */
         if (!otmp->oartifact && !otmp->cursed
             && otmp->otyp != WORM_TOOTH
-            && (otmp->spe + otmp->oerodeproof > (long) rnd(5)))
+            && ((long) otmp->spe + otmp->oerodeproof > (long) rnd(5)))
             otmp = weapon_oname(otmp);
         break;
     case FOOD_CLASS:
